Restructured outputs module for mass production
